# Improve null handling performance for nullable single input aggregation functions