--- a/pinot-core/src/main/java/org/apache/pinot/core/data/table/Record.java
+++ b/pinot-core/src/main/java/org/apache/pinot/core/data/table/Record.java
@@ -64,4 +64,9 @@ public class Record {
   public int hashCode() {
     return Arrays.hashCode(_values);
   }
+
+  @Override
+  public String toString() {
+    return Arrays.toString(_values);
+  }
 }