--- a/pinot-perf/src/main/java/org/apache/pinot/perf/SyntheticBlockValSets.java
+++ b/pinot-perf/src/main/java/org/apache/pinot/perf/SyntheticBlockValSets.java
@@ -240,7 +240,7 @@ public class SyntheticBlockValSets {
 
     @Override
     public FieldSpec.DataType getValueType() {
-      return FieldSpec.DataType.LONG;
+      return FieldSpec.DataType.DOUBLE;
     }
 
     @Override