@siddharthteotia 2024-08-13
I see a lot of code has changed on the critical path. Maybe it is just a lot of deletions / refactoring that is looking a lot. Do we have some performance number from a real production workload.

Btw, to be clear I am supportive of the change and the rationale behind it. But (if time permits) it may be useful to see results from real use case or at least some kind of CPU profiling numbers from FlameGraphs or JFR metrics even if it is on the synthetic data