@yashmayya 2024-08-28
Attaching some CPU profiles from local runs of the new sum query benchmark here:

[old-sum-nullhandling-false-nullperiod-128-summary-cpu.txt](https://github.com/user-attachments/files/16781021/old-sum-nullhandling-false-nullperiod-128-summary-cpu.txt)

[old-sum-nullhandling-true-nullperiod-128-summary-cpu.txt](https://github.com/user-attachments/files/16781022/old-sum-nullhandling-true-nullperiod-128-summary-cpu.txt)

[new-sum-nullhandling-false-nullperiod-128-summary-cpu.txt](https://github.com/user-attachments/files/16781024/new-sum-nullhandling-false-nullperiod-128-summary-cpu.txt)

[new-sum-nullhandling-true-nullperiod-128-summary-cpu.txt](https://github.com/user-attachments/files/16781025/new-sum-nullhandling-true-nullperiod-128-summary-cpu.txt)